Revert "Standardize competitionId and shooterId to Long throughout"
This reverts commit 4d4bb4f4c1759a3fb6fd9b2fc794319dbf259a83.

+++ app/src/main/java/se/kjellstrand/webshooter/data/results/ResultsRepository.kt
@@ -26,7 +26,7 @@ open class ResultsRepository @Inject constructor(
         private const val TAG = "ResultsRepository"
     }
 
-    fun get(competitionId: Long): Flow<Resource<ResultsResponse, UserError>> {
+    fun get(competitionId: Int): Flow<Resource<ResultsResponse, UserError>> {
         return flow {
             emit(Resource.Loading(true))
 
@@ -64,7 +64,7 @@ open class ResultsRepository @Inject constructor(
         }
     }
 
-    fun getPreferCached(competitionId: Long): Flow<Resource<ResultsResponse, UserError>> {
+    fun getPreferCached(competitionId: Int): Flow<Resource<ResultsResponse, UserError>> {
         return flow {
             emit(Resource.Loading(true))
             try {
@@ -99,7 +99,7 @@ open class ResultsRepository @Inject constructor(
         }
     }
 
-    fun getShooterResults(competitionId: Long, shooterId: Long): Flow<Resource<ResultsResponse, UserError>> {
+    fun getShooterResults(competitionId: Int, shooterId: Int): Flow<Resource<ResultsResponse, UserError>> {
         return flow {
             emit(Resource.Loading(true))
 
@@ -109,7 +109,7 @@ open class ResultsRepository @Inject constructor(
                 if (cached.isNotEmpty()) {
                     hasCached = true
                     val filtered = cached.map { it.toDomain(gson) }
-                        .filter { it.signup.user.userID == shooterId }
+                        .filter { it.signup.user.userID == shooterId.toLong() }
                     emit(Resource.Success(ResultsResponse(results = filtered)))
                     return@flow
                 }
@@ -136,7 +136,7 @@ open class ResultsRepository @Inject constructor(
             dao.deleteByCompetition(competitionId)
             dao.insertAll(result.results.map { it.toEntity(gson) })
 
-            val filtered = result.results.filter { it.signup.user.userID == shooterId }
+            val filtered = result.results.filter { it.signup.user.userID == shooterId.toLong() }
             emit(Resource.Success(result.copy(results = filtered)))
         }
     }
