+++ app/src/main/java/se/kjellstrand/webshooter/data/results/remote/ResultsRemoteDataSource.kt
@@ -1,12 +1,13 @@
 package se.kjellstrand.webshooter.data.results.remote
 
 import retrofit2.http.GET
+import retrofit2.http.POST
 import retrofit2.http.Path
 
 interface ResultsRemoteDataSource {
 
     @GET("/api/v4.1.9/competitions/{id}/results")
     suspend fun getResults(
-        @Path("id") id: Long,
+        @Path("id") id: Int,
     ): ResultsResponse
 }
